current data on multiple securities at a time
msGetBBG[] can now handle requests for multiple securities at a time, a much faster way to get results than the old way of leaving a LinkObject live and sending multiple requests through it (and even faster still when compared to the original method of launching a new link for each request). This completely obviates the old bbGetCurrent[], which survived in recent versions as bbGetMulti[], and this function has been removed.